requirements.txt
lxml
aiogram
pillow
pandas
requests
openpyxl
